Add preliminary Enemy class Rework HP & EXP bar to have its own ProgressBar class in Gui namespace

+++ Project Omega/Sword.cpp
@@ -30,8 +30,8 @@ void Sword::update(const sf::Vector2f mouse_pos_view, const sf::Vector2f center)
 	float dX = mouse_pos_view.x - this->weaponSprite.getPosition().x;
 	float dY = mouse_pos_view.y - this->weaponSprite.getPosition().y;
 
-	const float PI = static_cast<float>(3.14159265);
-	float deg = atan2(dY, dX) * 180 / PI;
+	const float PI = 3.14159265f;
+	float deg = atan2(dY, dX) * 180.f / PI;
 
 	this->weaponSprite.setRotation(deg + 90.f);
 }
